requirements.txt
adbc_driver_manager==0.10.0
adbc_driver_postgresql==0.10.0
adbc_driver_sqlite==0.10.0
AppKit==0.2.8
asgiref==3.7.2
attr==0.3.2
beautifulsoup4==4.12.3
boto3==1.34.50
botocore==1.34.50
brotli==1.1.0
brotlicffi==1.1.0.0
celery==5.3.6
coloredlogs==15.0.1
ConfigParser==6.0.1
contextlib2==21.6.0
cryptography==41.0.7
cycler==0.12.1
Cython==3.0.8
dash_testing_stub==0.0.2
dask==2024.2.1
diskcache==5.6.3
dl==0.1.0
docutils==0.20.1
fastparquet==2024.2.0
filelock==3.13.1
fire==0.5.0
flask_compress==1.14
Foundation==0.1.0a0.dev1
fsspec==2024.2.0
h2==4.1.0
HTMLParser==0.0.2
hypothesis==6.98.12
ipykernel==6.29.2
ipython==8.12.3
ipywidgets==8.1.2
jnius==1.1.0
js==1.0
keyring==24.3.0
lxml==5.1.0
matplotlib==3.8.3
mock==5.1.0
mtrand==0.1
multiprocess==0.70.16
mypy==1.8.0
numarray==1.5.1
Numeric==24.2
numexpr==2.9.0
odfpy==1.4.1
openpyxl==3.1.2
orjson==3.9.15
percy==2.0.2
pickle5==0.0.12
Pillow==9.5.0
Pillow==10.2.0
protobuf==4.25.3
psutil==5.9.8
py==1.11.0
PyInstaller==6.4.0
pyodide==0.0.2
pyOpenSSL==24.0.0
PyQt4==4.11.4
PyQt5==5.15.10
PyQt5_sip==12.13.0
pytest==8.0.2
python-dotenv==1.0.1
python_calamine==0.1.7
pywin32==306
pyxlsb==1.0.10
PyYAML==6.0.1
PyYAML==6.0.1
qtpy==2.4.1
railroad==0.5.0
redis==5.0.1
s3fs==2024.2.0
scikit_learn==1.4.1.post1
scipy==1.12.0
selenium==4.18.1
sets==0.3.2
simplejson==3.19.2
skimage==0.0
Sphinx==7.2.6
SQLAlchemy==2.0.27
statsmodels==0.14.1
streamlit==1.30.0
tables==3.9.2
thread==0.1.3
threadpoolctl==3.3.0
tornado==6.4
traitlets==5.14.1
trove_classifiers==2024.2.23
urllib3_secure_extra==0.1.0
vincent==0.4.4
watchdog==3.0.0
xarray==2024.2.0
xlrd==2.0.1
xlsxwriter==3.2.0
xmlrpclib==1.0.1
zstandard==0.22.0
~treamlit==1.29.0
